+++ android/src/main/kotlin/com/hello/curiosity/regeneration/android/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.hello.curiosity.regenranch.android
+package com.hello.curiosity.regeneration.android
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -8,7 +8,7 @@ import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.hello.curiosity.regenranch.Greeting
+import com.hello.curiosity.regeneration.Greeting
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
