+++ shared/src/commonMain/kotlin/com/hello/curiosity/regeneration/Platform.kt
@@ -1,0 +1,7 @@
+package com.hello.curiosity.regeneration
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
