--- shared/src/commonMain/kotlin/com/hello/curiosity/regenranch/Platform.kt
@@ -1,7 +1,0 @@
-package com.hello.curiosity.regenranch
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
